+++ src/main/java/com/supermartijn642/wormhole/screen/EnergyBarWidget.java
@@ -7,7 +7,7 @@ import com.supermartijn642.core.gui.widget.WidgetRenderContext;
 import com.supermartijn642.core.gui.widget.premade.AbstractButtonWidget;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 
 import java.util.function.Consumer;
 import java.util.function.Supplier;
@@ -17,7 +17,7 @@ import java.util.function.Supplier;
  */
 public class EnergyBarWidget extends AbstractButtonWidget {
 
-    public static final ResourceLocation BARS = ResourceLocation.fromNamespaceAndPath("wormhole", "gui/energy_bars");
+    public static final Identifier BARS = Identifier.fromNamespaceAndPath("wormhole", "gui/energy_bars");
 
     private final Supplier<Integer> energy, capacity;